requirements.txt
aapt==1.0.2
allure-pytest==2.12.0
allure-python-commons==2.12.0
appnope==0.1.4
asttokens==2.4.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
by==0.0.4
certifi==2022.12.7
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.1.0
execnet==1.9.0
executing==2.0.1
fastjsonschema==2.19.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
libusb1==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
openpyxl==3.1.0
outcome==1.2.0
packaging==23.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
pluggy==1.0.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.17.2
PySocks==1.7.1
pytest==7.2.1
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-rerunfailures==11.1.2
pytest-xdist==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyVirtualDisplay==3.0
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
selenium==4.11.2
self==2020.12.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
trio==0.22.0
trio-websocket==0.9.2
typing==3.7.4.3
urllib3==1.26.14
wcwidth==0.2.13
webencodings==0.5.1
wsproto==1.2.0
yarg==0.1.9
